Update game dependencies and incorporate a new font for the game
The commit primarily updates the dependencies required for the game in go.mod and go.sum. In addition, a new font is embedded in the game for the title, game over and press start elements. The update, embed and usage of the new font involve changes in game.go and player.go. The game performance stats are also embedded into the screen for review.